requirements.txt
fastapi
uvicorn
streamlit
supabase
pandas
groq
python-dotenv
plotly
python-multipart
